APP-ERP/requirements.txt
customtkinter>=5.2.2
pillow>=10.0.0
reportlab>=4.0.0